clear require cache only for changed files
